+++ src/rover_pkg/camera_servo_test.py
@@ -1,18 +1,18 @@
 #!/usr/bin/env python3
 
 import rospy
-from std_msgs.msg import Int8
+from std_msgs.msg import Float32
 
 class CameraTester:
     def __init__(self):
         rospy.init_node("camera_servo_tester", anonymous=True)
-        self.pub = rospy.Publisher("/base/gui/out/camera_servo", Int8, queue_size=10)
+        self.pub = rospy.Publisher("/base/gui/out/camera_servo", Float32, queue_size=10)
         self.rate = rospy.Rate(1)
 
     def run(self):
         while True:
             set_point = input("Enter motor setpoint as a value from 0 to 255: ")
-            msg = Int8()
+            msg = Float32()
             msg.data = int(set_point)
             self.pub.publish(msg)
